# Faster megular expressions