--- a/util.go
+++ b/util.go
@@ -129,6 +129,6 @@ func ForEach(m Multiaddr, cb func(c Component) bool) {
 }
 
 func (m Multiaddr) Match(p ...meg.Pattern) (bool, error) {
-	matcher := meg.PatternToMatchState(p...)
+	matcher := meg.PatternToMatcher(p...)
 	return meg.Match(matcher, m)
 }