@sukunrt 2025-02-13
It would be interesting to have a benchmark for a multiaddr to String and then using a stdlib string regex on it.