@sukunrt 2025-02-13
Okay this is much slower:
```
func BenchmarkIsWebRTCDirectMultiaddrString(b *testing.B) {
	addr := multiaddr.StringCast("/ip4/1.2.3.4/udp/1234/webrtc-direct/")

	b.ResetTimer()
	b.ReportAllocs()
	reg := regexp.MustCompile(`^/ip4/(?P<ip>.+)/udp/(?P<port>.+)/webrtc-direct(/certhash/[^/]+)*$`)
	for i := 0; i < b.N; i++ {
		addrS := addr.String()
		res := reg.FindStringSubmatch(addrS)
		if len(res) != 4 {
			b.Fatal("unexpected result", addrS, len(res))
		}
	}
}
```